clean up  : remote `useTx` `useWithdrawV1`
