+++ packages/mixer/src/hooks/withdraw/useWithdraw.tsx
@@ -1,198 +1,85 @@
-import { useGroupTree, useMerkleProvider, useMixerInfo } from '@webb-dapp/mixer/hooks';
-import { useLeavesEvents } from '@webb-dapp/mixer/hooks/merkle/useLeavesEvents';
-import { useTreeLeaves } from '@webb-dapp/mixer/hooks/merkle/useTreeLeaves';
-import { useAccounts, useCall } from '@webb-dapp/react-hooks';
-import { useTX } from '@webb-dapp/react-hooks/tx/useTX';
-import { LoggerService } from '@webb-tools/app-util';
-import { Note } from '@webb-tools/sdk-mixer';
-import { Block } from '@webb-tools/types/interfaces';
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useWebContext, WebbContentState, WithdrawState } from '@webb-dapp/react-environment/webb-context';
 
-import { decodeAddress } from '@polkadot/keyring';
-import { u8aToHex } from '@polkadot/util';
-
-const logger = LoggerService.get('Withdraw');
-
-export enum WithdrawState {
-  Canceled,
-  Ideal,
-
-  GeneratingZk,
-
-  SendingTransaction,
-
-  Done,
-  Faild,
-}
-
-export type WithdrawTXInfo = {
-  account: string;
-  note: Note;
+export type UseWithdrawV2Props = {
+  note: string;
+  recipient: string;
 };
+export type WithdrawErrors = {
+  error: string;
 
-export function useWithdraw(noteStr: string) {
-  const { init, initialized, merkle, restart, restarting } = useMerkleProvider();
+  validationError: {
+    note: string;
+    recipient: string;
+  };
+};
+export const useWithdraw = (params: UseWithdrawV2Props) => {
+  const [stage, setStage] = useState<WithdrawState>(WithdrawState.Ideal);
+  const { activeApi } = useWebContext();
 
-  useEffect(() => {
-    init();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-  const blockNumber = useCall<Block>('query.system.number', []);
-  const cancel = useRef(false);
-  const note = useMemo(() => {
-    try {
-      return noteStr ? Note.deserialize(noteStr) : undefined;
-    } catch (e) {
-      return undefined;
-    }
-  }, [noteStr]);
-
-  const [withdrawTo, _setWithdrawTo] = useState<string>('');
-  const [stage, _setStage] = useState(WithdrawState.Ideal);
-
-  const noteMixerGroupId = useMemo(() => note?.id, [note]);
-  const groupTreeWrapper = useGroupTree(noteMixerGroupId?.toString());
-  const mixerInfoWrapper = useMixerInfo(noteMixerGroupId?.toString());
-  const treeLeaves = useTreeLeaves(noteMixerGroupId);
-  const withdrawTxInfo = useMemo<WithdrawTXInfo | null>(() => {
-    if (!note || !withdrawTo) {
-      return null;
-    }
-    return {
-      account: withdrawTo,
-      note,
-    };
-  }, [withdrawTo, note]);
-
-  const [validationErrors, setValidationError] = useState({
-    note: ``,
-    withdrawTo: ``,
+  const [error, setError] = useState<WithdrawErrors>({
+    error: '',
+    validationError: {
+      note: '',
+      recipient: '',
+    },
   });
-  // restart the merkle tree on note change to undo mutations
+  const withdrawApi = useMemo(() => {
+    const withdraw = activeApi?.methods.mixer.withdraw;
+    if (!withdraw?.enabled) return null;
+    return withdraw.inner;
+  }, [activeApi]);
+
+  // hook events
   useEffect(() => {
-    if (note) {
-      restart();
-    }
-  }, [note]);
-
-  useEffect(() => {
-    setValidationError((prev) => ({
-      ...prev,
-      note: note ? '' : `Please make sure to enter a valid note`,
-      withdrawTo: withdrawTo ? '' : 'Invalid recipient',
-    }));
-  }, [withdrawTo, note]);
-
-  const accounts = useAccounts();
-
-  const cancelWithdraw = useCallback(async () => {
-    if (stage < WithdrawState.SendingTransaction) {
-      logger.info(`Canceling the withdraw ZK`);
-      cancel.current = true;
-      await restart();
-      _setStage(WithdrawState.Canceled);
-    }
-  }, [stage, restart]);
-
-  useEffect(() => {
-    setWithdrawTo(accounts.active?.address.toString() || '');
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [accounts.active?.address]);
-
-  const setWithdrawTo = useCallback((next: string) => {
-    _setWithdrawTo(next);
-  }, []);
-
-  const setNextStageWithThresholdToIdeal = useCallback((success: boolean) => {
-    _setStage(success ? WithdrawState.Done : WithdrawState.Faild);
-    setTimeout(() => {
-      _setStage(WithdrawState.Ideal);
-    }, 10000);
-  }, []);
-
-  const { executeTX, loading } = useTX({
-    method: 'withdraw',
-    onExtrinsicSuccess: () => setNextStageWithThresholdToIdeal(true),
-    onFailed: () => setNextStageWithThresholdToIdeal(false),
-    params: [],
-    section: 'mixer',
-  });
-
-  const canCancel = useMemo(() => {
-    return stage < WithdrawState.SendingTransaction && stage > WithdrawState.Ideal && !restarting;
-  }, [stage, restarting]);
-
-  const canWithdraw = useMemo(() => {
-    const validState = stage < WithdrawState.GeneratingZk;
-    return (
-      groupTreeWrapper.ready && mixerInfoWrapper.ready && Boolean(note) && initialized && validState && treeLeaves.done
-    );
-  }, [groupTreeWrapper, mixerInfoWrapper, note, initialized, stage, treeLeaves]);
-
-  const withdraw = async () => {
-    const root = groupTreeWrapper.rootHashU8a;
-    if (!root || !note) {
-      logger.error(`Root has Error`);
-      return;
-    }
-    if (!merkle) {
-      logger.error(`Mixer isn't initialized`);
-      return;
-    }
-    if (!groupTreeWrapper.ready || !mixerInfoWrapper.ready) {
-      logger.error(`Groups aren't ready`);
-      return;
-    }
-    if (!withdrawTo) {
-      logger.error(`withdrawTo isn't ready`);
-      return;
-    }
-    if (!treeLeaves.done) {
-      logger.error(`Still fetching leaves`);
-      return;
-    }
-    /* Generating Withdraw proof*/
-    _setStage(WithdrawState.GeneratingZk);
-    await merkle.addLeaves(treeLeaves.data, root);
-    const zk = await merkle.generateZKProof({
-      note: noteStr,
-      recipient: decodeAddress(withdrawTo),
-      relayer: decodeAddress(withdrawTo),
-      root,
+    const unsubscribe: Record<string, (() => void) | void> = {};
+    if (!withdrawApi) return;
+    unsubscribe['stateChange'] = withdrawApi.on('stateChange', (stage: WithdrawState) => {
+      setStage(stage);
     });
 
-    /* Generating Withdraw proof*/
+    unsubscribe['validationError'] = withdrawApi.on('validationError', (validationError) => {
+      setError((p) => ({
+        ...p,
+        validationError,
+      }));
+    });
+    unsubscribe['error'] = withdrawApi.on('error', (withdrawError) => {
+      setError((p) => ({
+        ...p,
+        error: withdrawError,
+      }));
+    });
 
-    const withdrawProof = {
-      cached_block: blockNumber,
-      cached_root: root,
-      comms: zk.commitments,
-      leaf_index_commitments: zk.leafIndexCommitments,
-      mixer_id: noteMixerGroupId,
-      nullifier_hash: zk.nullifierHash,
-      proof_bytes: u8aToHex(zk.proof),
-      proof_commitments: zk.proofCommitments,
-      recipient: withdrawTo,
-      relayer: withdrawTo,
-    };
-    if (!cancel.current) {
-      _setStage(WithdrawState.SendingTransaction);
-      await executeTX([withdrawProof]);
-    } else {
-      logger.info(`TX cancel`);
+    return () => Object.values(unsubscribe).forEach((v) => v && v());
+  }, [withdrawApi]);
+
+  const withdraw = useCallback(async () => {
+    if (!withdrawApi) return;
+
+    if (stage === WithdrawState.Ideal) {
+      const { note, recipient } = params;
+      console.log(params);
+      await withdrawApi.withdraw(note, recipient);
     }
-  };
+  }, [withdrawApi, params]);
+
+  const canCancel = useMemo(() => {
+    return stage < WithdrawState.SendingTransaction && stage > WithdrawState.Ideal;
+  }, [stage, withdrawApi]);
+
+  const cancelWithdraw = useCallback(async () => {
+    if (canCancel) {
+      await withdrawApi?.cancelWithdraw();
+    }
+  }, [canCancel]);
+
   return {
-    accounts: accounts.accounts || [],
-    canCancel,
-    canWithdraw,
-    cancelWithdraw,
-    loading,
-    setWithdrawTo,
     stage,
-    validationErrors,
     withdraw,
-    withdrawTo,
-    withdrawTxInfo,
+    canCancel,
+    cancelWithdraw,
+    error: error.error,
+    validationErrors: error.validationError,
   };
-}
+};
