--- packages/react-hooks/src/tx/index.ts
@@ -1,1 +1,0 @@
-export {};
